Create NestableTypeVisitor and associated spec
