+++ Sources/SwiftInspectorVisitors/NestableTypeInfo.swift
@@ -22,26 +22,38 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-public struct ClassInfo: Codable, Equatable {
-  public let name: String
-  public let inheritsFromTypes: [TypeDescription]
-  public let parentType: TypeDescription?
-  public let modifiers: Set<String>
-  // TODO: also find and expose properties on this declaration
+protocol NestableTypeInfo {
+
+  init(
+    name: String,
+    inheritsFromTypes: [TypeDescription],
+    parentType: TypeDescription?,
+    modifiers: Set<String>)
+
+  var name: String { get }
+  var inheritsFromTypes: [TypeDescription] { get }
+  var parentType: TypeDescription? { get }
+  var modifiers: Set<String> { get }
+  // TODO: also find and expose properties on this type
 }
 
-public struct EnumInfo: Codable, Equatable {
+public struct ClassInfo: NestableTypeInfo, Codable, Equatable {
   public let name: String
   public let inheritsFromTypes: [TypeDescription]
   public let parentType: TypeDescription?
   public let modifiers: Set<String>
-  // TODO: also find and expose properties on this declaration
 }
 
-public struct StructInfo: Codable, Equatable {
+public struct EnumInfo: NestableTypeInfo, Codable, Equatable {
   public let name: String
   public let inheritsFromTypes: [TypeDescription]
   public let parentType: TypeDescription?
   public let modifiers: Set<String>
-  // TODO: also find and expose properties on this declaration
+}
+
+public struct StructInfo: NestableTypeInfo, Codable, Equatable {
+  public let name: String
+  public let inheritsFromTypes: [TypeDescription]
+  public let parentType: TypeDescription?
+  public let modifiers: Set<String>
 }
